Change test time by clicking test time
